+++ node/src/types/kernel/api/block/getBlockBreadcrumb/payload.d.ts
@@ -23,7 +23,7 @@ export interface IPayload {
     /**
      * The block types that needs to be excluded
      */
-    excludeTypes?: ExcludeType[];
+    excludeTypes?: TBlockType[];
     /**
      * Block ID
      */
@@ -33,6 +33,6 @@ export interface IPayload {
 /**
  * Block type
  */
-export type ExcludeType = "NodeDocument" | "NodeSuperBlock" | "NodeBlockquote" | "NodeList" | "NodeListItem" | "NodeHeading" | "NodeParagraph" | "NodeMathBlock" | "NodeTable" | "NodeCodeBlock" | "NodeHTMLBlock" | "NodeThematicBreak" | "NodeAudio" | "NodeVideo" | "NodeIFrame" | "NodeWidget" | "NodeBlockQueryEmbed";
+export type TBlockType = "NodeDocument" | "NodeSuperBlock" | "NodeBlockquote" | "NodeList" | "NodeListItem" | "NodeHeading" | "NodeParagraph" | "NodeMathBlock" | "NodeTable" | "NodeCodeBlock" | "NodeHTMLBlock" | "NodeThematicBreak" | "NodeAudio" | "NodeVideo" | "NodeIFrame" | "NodeWidget" | "NodeBlockQueryEmbed" | "NodeAttributeView";
 
 //#endregion content